Move TeamMetadataForm to teams.forms for reuse
Both the per-team settings page and the new admin team-detail page need it.
